+++ src/algorithms/joint_marginals.jl
