+++ test/engines/test_free_energy_assemblers.jl
@@ -1,0 +1,11 @@
+module MessagePassingAssemblersTest
+
+using Test
+using ForneyLab
+import ForneyLab:
+
+@testset "..." begin
+	@test true == false
+end
+
+end # module
